refactor: Clean up the blender-openjd CLI interface
* Modify __main__.py to follow the guidelines on `python -m`
  entry points by not having 'if __name__ == "__main__"'.
* Add blender_adaptor/BlenderAdaptor/_cli.py to hold the main()
  function for the CLI interface.
