+++ src/deadline/blender_adaptor/BlenderAdaptor/__init__.py
@@ -1,6 +1,6 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
-from .__main__ import main
+from ._cli import main
 from .adaptor import BlenderAdaptor
 
 __all__ = [
